Unify events interfaces/apis; need to figure out visibility next
