--- source/sdk/src/main/java/com/stytch/sdk/b2b/events/Events.kt
@@ -1,9 +1,0 @@
-package com.stytch.sdk.b2b.events
-
-internal interface Events {
-    fun logEvent(
-        eventName: String,
-        details: Map<String, Any>? = null,
-        error: Exception? = null,
-    )
-}
