+++ source/sdk/src/main/java/com/stytch/sdk/b2b/network/StytchB2BApi.kt
@@ -61,6 +61,7 @@ import com.stytch.sdk.common.WEB_URL
 import com.stytch.sdk.common.dfp.CaptchaProvider
 import com.stytch.sdk.common.dfp.DFPProvider
 import com.stytch.sdk.common.errors.StytchSDKNotConfiguredError
+import com.stytch.sdk.common.events.EventsAPI
 import com.stytch.sdk.common.network.ApiService
 import com.stytch.sdk.common.network.InfoHeaderModel
 import com.stytch.sdk.common.network.StytchAuthHeaderInterceptor
@@ -775,8 +776,8 @@ internal object StytchB2BApi {
             }
     }
 
-    internal object Events {
-        suspend fun logEvent(
+    internal object Events : EventsAPI {
+        override suspend fun logEvent(
             eventId: String,
             appSessionId: String,
             persistentId: String,
@@ -784,8 +785,8 @@ internal object StytchB2BApi {
             timezone: String,
             eventName: String,
             infoHeaderModel: InfoHeaderModel,
-            details: Map<String, Any>? = null,
-            error: Exception? = null,
+            details: Map<String, Any>?,
+            error: Exception?,
         ): NoResponseResponse =
             safeB2BApiCall {
                 apiService.logEvent(
